Add support for whitelisting shader attributes so that, in case they can't be bound to valid vertex data, they're not reported to the application as errors.
This kind of validation did not exist in pre-RenderCore times. cocos3d will now render objects with bad vertex data (e.g. meshes that don't have all necessary attributes) with a placeholder shader. For dragons3d, we managed to fix all such issues in shader or art, but tow has considerably more instances of this problem and it would require dozens, if note hundreds, of individual meshes to be re-exported (and that's considering the vertex data even exists in Maya).

We could simply disable all such validation, but it's a useful tool to have available in the future. Instead, we whitelist attributes that are known to be problematic on a per-project basis. The dream is to one day fix all such occurrences and remove this code, but that remains to be seen.

Bonus:
- Expose the shader source identifiers in ShaderProgram.
- Pretty formatting of IteratorRange objects in Log statements.