Make SyncEventSet unit test pass reliably
On GL, we wait until we're caught up before checking the sync event
set. On other targets, we were doing nothing, so we passed only if
things are going slowly.

Unfortunately, we don't really have a cross-target API to call
here. `SyncEventSet.Stall()` should work, but it could also (depending
on the implementation) make us pass even when we shouldn't be
passing.) So, I added that as a fallback, but also added an Apple
Metal specific wait.