Simple improvements to many cmake files
Expanded most of the GLOB() calls to an explicit list of files. This gets around issues where the glob is not automatically recalculated when the filesystem changes. Also made the include folders from the foreign libraries part of the ForeignMisc project interface -- so now projects can just link ForeignMisc to get those include folders, instead of having to repeat them everywhere.